test: Fix cleanup on v2
