+++ tests/integration/dbapi/async/V2/test_queries_async.py
@@ -2,10 +2,10 @@ import math
 from datetime import date, datetime
 from decimal import Decimal
 from os import environ
-from random import choice, randint
+from random import randint
 from typing import Callable, Generator, List
 
-from pytest import fixture, mark, raises
+from pytest import mark, raises
 
 from firebolt.async_db import Binary, Connection, Cursor, OperationalError
 from firebolt.async_db.connection import connect
@@ -432,17 +432,6 @@ async def test_bytea_roundtrip(
         assert (
             bytes_data.decode("utf-8") == data
         ), "Invalid bytea data returned after roundtrip"
-
-
-@fixture
-async def setup_v2_db(connection_system_engine_v2: Connection, use_db_name: str):
-    use_db_name = use_db_name + "_async"
-    with connection_system_engine_v2.cursor() as cursor:
-        # randomize the db name to avoid conflicts
-        suffix = "".join(choice("0123456789") for _ in range(2))
-        await cursor.execute(f"CREATE DATABASE {use_db_name}{suffix}")
-        yield f"{use_db_name}{suffix}"
-        await cursor.execute(f"DROP DATABASE {use_db_name}{suffix}")
 
 
 @mark.xfail("dev" not in environ[API_ENDPOINT_ENV], reason="Only works on dev")
