+++ tests/integration/dbapi/sync/V2/test_system_engine.py
@@ -83,18 +83,17 @@ def test_system_engine_v2_account(connection_system_engine_v2: Connection):
 
 
 def test_system_engine_use_engine(
-    connection_system_engine_v2: Connection, database_name: str, engine_name: str
+    connection_system_engine_v2: Connection, setup_v2_db: str, engine_v2: str
 ):
+    table_name = "test_table_sync"
     with connection_system_engine_v2.cursor() as cursor:
-        cursor.execute(f"CREATE DATABASE IF NOT EXISTS {database_name}")
-        cursor.execute(f"USE DATABASE {database_name}")
-        cursor.execute(f"CREATE ENGINE IF NOT EXISTS {engine_name}")
-        cursor.execute(f"USE ENGINE {engine_name}")
-        cursor.execute("CREATE TABLE IF NOT EXISTS test_table (id int)")
+        cursor.execute(f"USE DATABASE {setup_v2_db}")
+        cursor.execute(f"USE ENGINE {engine_v2}")
+        cursor.execute(f"CREATE TABLE IF NOT EXISTS {table_name} (id int)")
         # This query fails if we're not on a user engine
-        cursor.execute("INSERT INTO test_table VALUES (1)")
+        cursor.execute(f"INSERT INTO {table_name} VALUES (1)")
         cursor.execute("USE ENGINE system")
         # Werify we've switched to system by making previous query fail
         with raises(OperationalError):
-            cursor.execute("INSERT INTO test_table VALUES (1)")
-        cursor.execute("DROP TABLE IF EXISTS test_table")
+            cursor.execute(f"INSERT INTO {table_name} VALUES (1)")
+        cursor.execute(f"DROP TABLE IF EXISTS {table_name}")
